// src/LuckyMaze.Frontend/src/app/api/model/mazeSize.ts
/**
 * LuckyMaze.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const MazeSize = {
    Small16x16: 'Small16x16',
    Grid21x21: 'Grid21x21',
    Medium32x32: 'Medium32x32',
    Large64x64: 'Large64x64'
} as const;
export type MazeSize = typeof MazeSize[keyof typeof MazeSize];